+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\src",
+    "\\src\\HabitsMS",
+    "\\src\\HabitsMS\\src"
+  ],
+  "SelectedNode": "\\src\\HabitsMS\\HabitsMS.sln",
+  "PreviewInSolutionExplorer": false
+}
