+++ src/HabitsMS/.gitignore
@@ -1,0 +1,15 @@
+.idea
+.DS_Store
+src/.DS_Store
+src/Core/bin
+src/Core/obj
+src/Core/.DS_Store
+src/Networking.API/bin
+src/Networking.API/obj
+src/Networking.API/.DS_Store
+src/Store.PostgreSQL/bin
+src/Store.PostgreSQL/obj
+src/Store.PostgreSQL/.DS_Store
+src/HabitsMS.API/bin
+src/HabitsMS.API/obj
+src/HabitsMS.API/.DS_Store
