+++ src/HabitsMS/src/HabitsMS.API/HabitsMS.API.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+        <RootNamespace>HabitsMS.API</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Networking.API\Networking.API.csproj" />
+      <ProjectReference Include="..\Store.PostgreSQL\Store.PostgreSQL.csproj" />
+    </ItemGroup>
+
+</Project>
