+++ src/HabitsMS/src/HabitsMS.API/Program.cs
@@ -1,0 +1,11 @@
+﻿using Networking.API;
+using Store.PostgreSQL.Repositories;
+using Core.Stores;
+using Microsoft.Extensions.DependencyInjection;
+
+var api = new ApiBootstrap(args, options =>
+{
+    options.AddScoped<IHabitStore, HabitRepository>();
+});
+
+await api.RunAsync();
