+++ src/HabitsMS/src/Networking.API/Networking.API.csproj
