--- src/UserMS/.gitignore
@@ -1,15 +1,0 @@
-.idea
-.DS_Store
-src/.DS_Store
-src/Core/bin
-src/Core/obj
-src/Core/.DS_Store
-src/Networking.API/bin
-src/Networking.API/obj
-src/Networking.API/.DS_Store
-src/Store.PostgreSQL/bin
-src/Store.PostgreSQL/obj
-src/Store.PostgreSQL/.DS_Store
-src/UserMS.API/bin
-src/UserMS.API/obj
-src/UserMS.API/.DS_Store
