--- src/UserMS/UserMS.sln
@@ -1,42 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{159317CF-11A2-48B6-A7FA-98301923A3C0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "src\Core\Core.csproj", "{F74778FE-851C-4E49-97A5-D280C5C59BD8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Networking.API", "src\Networking.API\Networking.API.csproj", "{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Store.PostgreSQL", "src\Store.PostgreSQL\Store.PostgreSQL.csproj", "{60500477-3E8C-46B3-8369-2716691D925C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserMS.API", "src\UserMS.API\UserMS.API.csproj", "{A3000C8F-BDE4-4B97-B108-8B212E198D94}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F74778FE-851C-4E49-97A5-D280C5C59BD8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F74778FE-851C-4E49-97A5-D280C5C59BD8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F74778FE-851C-4E49-97A5-D280C5C59BD8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F74778FE-851C-4E49-97A5-D280C5C59BD8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62}.Release|Any CPU.Build.0 = Release|Any CPU
-		{60500477-3E8C-46B3-8369-2716691D925C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60500477-3E8C-46B3-8369-2716691D925C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60500477-3E8C-46B3-8369-2716691D925C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60500477-3E8C-46B3-8369-2716691D925C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A3000C8F-BDE4-4B97-B108-8B212E198D94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A3000C8F-BDE4-4B97-B108-8B212E198D94}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A3000C8F-BDE4-4B97-B108-8B212E198D94}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A3000C8F-BDE4-4B97-B108-8B212E198D94}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{F74778FE-851C-4E49-97A5-D280C5C59BD8} = {159317CF-11A2-48B6-A7FA-98301923A3C0}
-		{74B0C0E7-9412-47C9-8D8E-1D3EBEF89C62} = {159317CF-11A2-48B6-A7FA-98301923A3C0}
-		{60500477-3E8C-46B3-8369-2716691D925C} = {159317CF-11A2-48B6-A7FA-98301923A3C0}
-		{A3000C8F-BDE4-4B97-B108-8B212E198D94} = {159317CF-11A2-48B6-A7FA-98301923A3C0}
-	EndGlobalSection
-EndGlobal
