--- src/UserMS/src/Core/Entities/Interfaces/ISoftDeletable.cs
@@ -1,6 +1,0 @@
-namespace Core.Entities.Interfaces;
-
-public interface ISoftDeletable
-{
-    public DateTime? DeletedOn { get; set; }
-}
