--- src/UserMS/src/Core/Exceptions/WeakPasswordException.cs
@@ -1,18 +1,0 @@
-namespace Core.Exceptions;
-
-public class WeakPasswordException : Exception
-{
-    public WeakPasswordException()
-    {
-    }
-
-    public WeakPasswordException(string message)
-        : base(message)
-    {
-    }
-
-    public WeakPasswordException(string message, Exception inner)
-        : base(message, inner)
-    {
-    }
-}
