--- src/UserMS/src/Core/Stores/IUserStore.cs
@@ -1,9 +1,0 @@
-using Core.Entities;
-
-namespace Core.Stores;
-
-public interface IUserStore
-{
-    public Task<int> CreateUser(User user);
-    public Task<User> GetUser();
-}
