--- src/UserMS/src/Store.PostgreSQL/Database/Context.cs
@@ -1,21 +1,0 @@
-using Core.Entities;
-using Core.Entities.Interfaces;
-using Microsoft.EntityFrameworkCore;
-using Store.PostgreSQL.Extensions;
-
-namespace Store.PostgreSQL.Database;
-
-public class Context : DbContext
-{
-    public Context(DbContextOptions options) : base(options) {}
-    
-    public DbSet<User> Users { get; set; }
-
-    public DbSet<PasswordReset> PasswordResets { get; set; }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        modelBuilder.ApplyGlobalFilters<ISoftDeletable>(p => p.DeletedOn == null);
-        base.OnModelCreating(modelBuilder);
-    }
-}
