--- src/UserMS/src/Store.PostgreSQL/appsettings.Development.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "UserMSDatabase": "User ID=userms;Password=planner;Host=localhost;Port=5401;Database=userms;Pooling=true;"
-  }
-}
