--- src/UserMS/src/Store.PostgreSQL/appsettings.Staging.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "UserMSDatabase": "User ID=userms;Password=planner;Host=usermspgsql;Port=5432;Database=userms;Pooling=true;"
-  }
-}
