--- src/UserMS/src/Store.PostgreSQL/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "UserMSDatabase": ""
-  },
-  "AllowedHosts": "*"
-}
